+++ test/sidecar_test.exs
@@ -155,11 +155,12 @@ defmodule SidecarTest do
     Task.async(fn ->
       NewRelic.start_transaction("Test", "manual_connection")
       NewRelic.add_attributes(foo1: "BAR")
-
       pid = self()
 
+      tx = NewRelic.get_transaction()
+
       spawn(fn ->
-        NewRelic.connect_to_transaction(pid)
+        NewRelic.connect_to_transaction(tx)
         NewRelic.add_attributes(foo2: "BAZ")
 
         NewRelic.disconnect_from_transaction()
@@ -180,9 +181,6 @@ defmodule SidecarTest do
       refute attributes[:foo3]
     end)
     |> Task.await()
-
-    # no-op when called outside a Transaction
-    NewRelic.connect_to_transaction(self())
   end
 
   defmodule Traced do
@@ -196,6 +194,17 @@ defmodule SidecarTest do
     @trace :hello
     def hello do
       :hello
+    end
+
+    @trace :connected_task_async_nolink
+    def connected_task_async_nolink(fun) do
+      tx = NewRelic.get_transaction()
+
+      Task.Supervisor.async_nolink(TestTaskSup, fn ->
+        NewRelic.connect_to_transaction(tx)
+        fun.()
+      end)
+      |> Task.await()
     end
   end
 
@@ -212,31 +221,27 @@ defmodule SidecarTest do
         TestTaskSup,
         fn ->
           NewRelic.add_attributes(async_nolink: :NO)
-          Traced.hey()
         end
       )
       |> Task.await()
 
-      Task.Supervisor.async_nolink(
-        TestTaskSup,
-        fn ->
-          NewRelic.connect_task_to_transaction()
-          NewRelic.add_attributes(async_nolink_connected: :YES)
-          Traced.hey()
+      Traced.connected_task_async_nolink(fn ->
+        NewRelic.add_attributes(async_nolink_connected: :YES)
+        Traced.hey()
 
-          Task.Supervisor.async_stream_nolink(
-            TestTaskSup,
-            [:YES],
-            fn val ->
-              NewRelic.connect_task_to_transaction()
-              NewRelic.add_attributes(async_stream_nolink_connected: val)
-              Traced.hello()
-            end
-          )
-          |> Enum.map(& &1)
-        end
-      )
-      |> Task.await()
+        tx = NewRelic.get_transaction()
+
+        Task.Supervisor.async_stream_nolink(
+          TestTaskSup,
+          [:YES],
+          fn val ->
+            NewRelic.connect_to_transaction(tx)
+            NewRelic.add_attributes(async_stream_nolink_connected: val)
+            Traced.hello()
+          end
+        )
+        |> Enum.map(& &1)
+      end)
 
       send(test, :done)
     end)
@@ -262,23 +267,29 @@ defmodule SidecarTest do
         attr[:name] == "Transaction Root Process"
       end)
 
-    connected_task_process_span =
+    task_triggering_function_span =
       Enum.find(spans, fn %{attributes: attr} ->
-        attr[:"parent.id"] == tx_root_process_span[:id]
+        attr[:name] == "SidecarTest.Traced.connected_task_async_nolink/1"
       end)
 
-    assert connected_task_process_span.attributes[:name] == "Process"
+    assert task_triggering_function_span.attributes[:"parent.id"] == tx_root_process_span[:id]
+
+    task_triggered_process_span =
+      Enum.find(spans, fn %{attributes: attr} ->
+        attr[:name] == "Process" &&
+          attr[:"parent.id"] == task_triggering_function_span[:id]
+      end)
 
     hey_function_span =
       Enum.find(spans, fn %{attributes: attr} ->
         attr[:name] == "SidecarTest.Traced.hey/0"
       end)
 
-    assert hey_function_span.attributes[:"parent.id"] == connected_task_process_span[:id]
+    assert hey_function_span.attributes[:"parent.id"] == task_triggered_process_span[:id]
 
     connected_stream_task_process_span =
       Enum.find(spans, fn %{attributes: attr} ->
-        attr[:"parent.id"] == connected_task_process_span[:id]
+        attr[:"parent.id"] == task_triggered_process_span[:id]
       end)
 
     assert connected_stream_task_process_span.attributes[:name] == "Process"
@@ -289,8 +300,57 @@ defmodule SidecarTest do
       end)
 
     assert hello_function_span.attributes[:"parent.id"] == connected_stream_task_process_span[:id]
+  end
 
-    # no-op when called outside a Transaction
-    NewRelic.connect_task_to_transaction()
+  test "No-op when manually connecting inside an existing Transaction" do
+    TestHelper.restart_harvest_cycle(TelemetrySdk.Spans.HarvestCycle)
+    test = self()
+
+    t1 =
+      Task.async(fn ->
+        NewRelic.start_transaction("Test", "first_tx")
+
+        send(test, {:tx, NewRelic.get_transaction()})
+
+        receive do
+          :finish -> :ok
+        end
+      end)
+
+    tx =
+      receive do
+        {:tx, tx} -> tx
+      end
+
+    Task.async(fn ->
+      NewRelic.start_transaction("Test", "double_connect_test")
+      NewRelic.add_attributes(root: :YES)
+
+      Task.async(fn ->
+        # Should stay connected to parent Transaction:
+        NewRelic.connect_to_transaction(tx)
+
+        NewRelic.add_attributes(async_nolink_connected: :YES)
+      end)
+      |> Task.await()
+    end)
+    |> Task.await()
+
+    send(t1.pid, :finish)
+
+    [%{spans: spans}] = TestHelper.gather_harvest(TelemetrySdk.Spans.Harvester)
+
+    spansaction =
+      Enum.find(spans, fn %{attributes: attr} ->
+        attr[:category] == "Transaction" && attr[:name] == "Test/double_connect_test"
+      end)
+
+    assert spansaction.attributes[:root] == "YES"
+    assert spansaction.attributes[:async_nolink_connected] == "YES"
+  end
+
+  test "no-op when attempting to connect outside a Transaction" do
+    tx = NewRelic.get_transaction()
+    NewRelic.connect_to_transaction(tx)
   end
 end